[nasa/cryptolib#145] Created sa folder with both internal and mariadb options;
