+++ src/CMakeLists.txt
@@ -19,11 +19,19 @@ include_directories(../include)
 
 aux_source_directory(src_main LIB_SRC_FILES)
 if(MYSQL)
-    aux_source_directory(src_mysql LIB_SRC_MYSQL_FILES)
-    list(APPEND LIB_SRC_FILES ${LIB_SRC_MYSQL_FILES})
+    aux_source_directory(sa/mariadb MARIADB_FILES)
+    list(APPEND LIB_SRC_FILES ${MARIADB_FILES})
 else()
-    aux_source_directory(src_feature_stubs/mysql_stub LIB_SRC_MYSQL_FILES)
-    list(APPEND LIB_SRC_FILES ${LIB_SRC_MYSQL_FILES})
+    aux_source_directory(sa/mariadb_stub MARIADB_FILES)
+    list(APPEND LIB_SRC_FILES ${MARIADB_FILES})
+endif()
+
+if(SA_INTERNAL)
+    aux_source_directory(sa/internal SA_INTERNAL_FILES)
+    list(APPEND LIB_SRC_FILES ${SA_INTERNAL_FILES})
+else()
+    aux_source_directory(sa/internal_stub SA_INTERNAL_FILES)
+    list(APPEND LIB_SRC_FILES ${SA_INTERNAL_FILES})
 endif()
 
 if(LIBGCRYPT)
@@ -92,8 +100,8 @@ ELSE()
 ENDIF()
 
 IF(MYSQL)
-    file(GLOB MYSQL_SCRIPTS crypto_sadb/sadb_mariadb_sql/*.sql)
-    file(GLOB MYSQL_TEST_SCRIPTS crypto_sadb/test_sadb_mariadb_sql/*.sql)
+    file(GLOB MYSQL_SCRIPTS sa/sadb_mariadb_sql/*.sql)
+    file(GLOB MYSQL_TEST_SCRIPTS sa/test_sadb_mariadb_sql/*.sql)
     install(FILES ${MYSQL_SCRIPTS}
             DESTINATION ${CMAKE_INSTALL_PREFIX}/etc/sadb_mariadb_sql)
     install(FILES ${MYSQL_TEST_SCRIPTS}
