+++ src/sa/internal/sadb_routine_inmemory.template.c
