+++ src/sa/mariadb_stub/sadb_routine_mariadb.stub.c
