+++ src/sa/sadb_mariadb_sql/delete_sadb.sql
