+++ src/sa/sadb_mariadb_sql/empty_sadb.sql
